Fix bug when there is no data in scraper

+++ src/001_collect_data.py
@@ -25,40 +25,44 @@ def oil_scraper(year, month):
     calendar_week = soup.find_all('div', {'class': 'calendar week'})
     calendar_week_elements = soup.find_all('div', class_='cell border')
 
-    # Prepare the dataframe
-    df=pd.DataFrame(columns=["Week", "Unleaded 95", "Diesel", "LPG"])
-
-    # Scraper
-    column_names = list(df.columns.values.tolist())
-    columns = len(df.columns)
-    rows = int(len(calendar_week_elements)/columns)
-
-    i=0
-    for row in range(rows):
-        for column in range(columns):
-            if column_names[column] == "Week":
-                df.at[row, column_names[column]] = re.sub(re.compile(r'^[^0-9]*'), '', calendar_week_elements[i].text).strip()
-                i=i+1
-            else:
-                df.at[row, column_names[column]] = re.sub(re.compile(r'^[^0-9]*'), '', calendar_week_elements[i].text)[:-5]
-                i=i+1
+    if calendar_week_elements == []: 
+        return "false"
         
-    # Add end day
-    for row in range(rows):
-        df.at[row,'end_day'] = datetime.strptime(df.at[row, "Week"][-8:],'%d.%m.%y')
-    
-    # Add start day
-    for row in range(rows):
-        df.at[row,'start_day'] = datetime.strptime(df.at[row, "Week"][-8:],'%d.%m.%y') -  timedelta(days=6)
+    else:
+        # Prepare the dataframe
+        df=pd.DataFrame(columns=["Week", "Unleaded 95", "Diesel", "LPG"])
 
-    # Remove week columns
-    df = df.iloc[: , 1:]
+        # Scraper
+        column_names = list(df.columns.values.tolist())
+        columns = len(df.columns)
+        rows = int(len(calendar_week_elements)/columns)
 
-    # Reorganize columns
-    df = df[['start_day', 'end_day', 'Unleaded 95', 'Diesel', 'LPG']]
+        i=0
+        for row in range(rows):
+            for column in range(columns):
+                if column_names[column] == "Week":
+                    df.at[row, column_names[column]] = re.sub(re.compile(r'^[^0-9]*'), '', calendar_week_elements[i].text).strip()
+                    i=i+1
+                else:
+                    df.at[row, column_names[column]] = re.sub(re.compile(r'^[^0-9]*'), '', calendar_week_elements[i].text)[:-5]
+                    i=i+1
+            
+        # Add end day
+        for row in range(rows):
+            df.at[row,'end_day'] = datetime.strptime(df.at[row, "Week"][-8:],'%d.%m.%y')
+        
+        # Add start day
+        for row in range(rows):
+            df.at[row,'start_day'] = datetime.strptime(df.at[row, "Week"][-8:],'%d.%m.%y') -  timedelta(days=6)
+
+        # Remove week columns
+        df = df.iloc[: , 1:]
+
+        # Reorganize columns
+        df = df[['start_day', 'end_day', 'Unleaded 95', 'Diesel', 'LPG']]
 
 
-    return df
+        return df
 
 def upload_s3(bucket, new_data):
     s3 = boto3.client('s3')
@@ -91,20 +95,21 @@ def merge_datasets_S3():
 # Scraper
 currentYear = datetime.now().year
 currentMonth = datetime.now().month
-dataset = oil_scraper(currentYear, 2)
+dataset = oil_scraper(currentYear, currentMonth)
 
-# Upload S3
-upload_s3("gas-prices-project", dataset)
+if dataset != "false":
+    # Upload S3
+    upload_s3("gas-prices-project", dataset)
 
-# Merge
-concat_data = merge_datasets_S3()
+    # Merge
+    concat_data = merge_datasets_S3()
 
-# Upload S3
-upload_s3("gas-prices-project", concat_data)
+    # Upload S3
+    upload_s3("gas-prices-project", concat_data)
 
-# Rename
-s3 = boto3.resource('s3')
-s3.Object('gas-prices-project','data.csv').delete()
-s3.Object('gas-prices-project','data.csv').copy_from(CopySource='gas-prices-project/new_data.csv')
-s3.Object('gas-prices-project','new_data.csv').delete()
+    # Rename
+    s3 = boto3.resource('s3')
+    s3.Object('gas-prices-project','data.csv').delete()
+    s3.Object('gas-prices-project','data.csv').copy_from(CopySource='gas-prices-project/new_data.csv')
+    s3.Object('gas-prices-project','new_data.csv').delete()
